+++ ingenialink/ethercat/servo.py
@@ -1,7 +1,15 @@
-from typing import Optional
+from typing import Optional, TYPE_CHECKING
 
-from pysoem import CdefSlave, SdoError, MailboxError, PacketError, Emergency
 import ingenialogger
+
+try:
+    import pysoem
+except ImportError as ex:
+    pysoem = None
+    pysoem_import_error = ex
+
+if TYPE_CHECKING:
+    from pysoem import CdefSlave
 
 from ingenialink.exceptions import ILIOError
 from ingenialink.servo import Servo
@@ -22,6 +30,9 @@ class EthercatServo(Servo):
         dictionary_path: Path to the dictionary.
         servo_status_listener: Toggle the listener of the servo for
             its status, errors, faults, etc.
+
+    Raises:
+        ImportError: WinPcap is not installed
 
     """
 
@@ -52,11 +63,13 @@ class EthercatServo(Servo):
 
     def __init__(
         self,
-        slave: CdefSlave,
+        slave: "CdefSlave",
         slave_id: int,
         dictionary_path: str,
         servo_status_listener: bool = False,
     ):
+        if not pysoem:
+            raise pysoem_import_error
         self.__slave = slave
         self.slave_id = slave_id
         super(EthercatServo, self).__init__(slave_id, dictionary_path, servo_status_listener)
@@ -68,9 +81,9 @@ class EthercatServo(Servo):
         try:
             value: bytes = self.__slave.sdo_read(reg.idx, reg.subidx, buffer_size, complete_access)
             self._check_working_counter()
-        except (SdoError, MailboxError, PacketError, ILIOError) as e:
+        except (pysoem.SdoError, pysoem.MailboxError, pysoem.PacketError, ILIOError) as e:
             raise ILIOError(f"Error reading {reg.identifier}. Reason: {e}") from e
-        except Emergency as e:
+        except pysoem.Emergency as e:
             error_description = self._get_emergency_description(e.error_code)
             if error_description is None:
                 error_description = e
@@ -84,9 +97,9 @@ class EthercatServo(Servo):
         try:
             self.__slave.sdo_write(reg.idx, reg.subidx, data, complete_access)
             self._check_working_counter()
-        except (SdoError, MailboxError, PacketError, ILIOError) as e:
+        except (pysoem.SdoError, pysoem.MailboxError, pysoem.PacketError, ILIOError) as e:
             raise ILIOError(f"Error writing {reg.identifier}. Reason: {e}") from e
-        except Emergency as e:
+        except pysoem.Emergency as e:
             error_description = self._get_emergency_description(e.error_code)
             if error_description is None:
                 error_description = e
@@ -159,6 +172,6 @@ class EthercatServo(Servo):
         return error_description
 
     @property
-    def slave(self) -> CdefSlave:
+    def slave(self) -> "CdefSlave":
         """Ethercat slave"""
         return self.__slave